cast to float32

+++ tud_rl/iPDP_helper/feature_importance.py
@@ -2,7 +2,7 @@ import math
 import numpy as np
 
 
-def calculate_feature_importance(pdp_x: dict, pdp_y: dict) -> float:
+def calculate_feature_importance(pdp_x: dict, pdp_y: dict) -> np.float32:
     grid_points_x, grid_points_y = pdp_x.values(), pdp_y.values()
     mean_centered_pd = np.mean(list(grid_points_y))
     sum_individual_distance = 0
@@ -11,4 +11,5 @@ def calculate_feature_importance(pdp_x: dict, pdp_y: dict) -> float:
         mean_point_pd = grid_point_y - mean_centered_pd
         sum_individual_distance += mean_point_pd**2
 
-    return math.sqrt(1 / (n - 1) * sum_individual_distance)
+    # standard deviation for samples casted to float32
+    return np.float32(math.sqrt(1 / (n - 1) * sum_individual_distance))
